Move ConstraintVar to r1cs-core
